+++ Mlib/Scene/Audio/Engine_Audio.hpp
@@ -23,7 +23,7 @@ public:
         float tires_angular_velocity,
         const EnginePowerIntent& engine_power_intent,
         float max_surface_power) override;
-    virtual void set_position(const FixedArray<float, 3>& position) override;
+    virtual void set_position(const FixedArray<double, 3>& position) override;
 private:
 #ifndef WITHOUT_ALUT
     std::shared_ptr<AudioBufferSequence> driving_buffer_sequence_;
